pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "livekit",
    "livekit-agents[openai]",
    "livekit-plugins-openai",
    "livekit-agents[voiceassistant]",
    "livekit-plugins-noise-cancellation",
    "langchain",
    "langchain-openai",
    "langchain-chroma",
    "chromadb",
    "openai",
    "python-multipart",
    "aiofiles",
    "python-dotenv",
    "requests>=2.32.5",
    "sqlalchemy",
    "pydantic",
    "pydantic-settings",
    "tavily-python",
    "pypdf2",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "PyJWT",
    "pandas",
    "openpyxl",
    "winloop>=0.2.2",
]
